+++ src/main/java/com/congreso/backend/controller/SubmenuC.java
@@ -6,6 +6,6 @@ import org.springframework.http.ResponseEntity;
 public interface SubmenuC {
 
     public ResponseEntity<ApiResponse> findAll();
-    public ResponseEntity<ApiResponse> findById(long id_subm);
+    public ResponseEntity<ApiResponse> findById(int id_subm);
 
 }
