+++ src/main/java/com/congreso/backend/repositoryE/SubmenuRepo.java
@@ -4,5 +4,5 @@ import com.congreso.backend.entities.PrivilegiosE;
 import com.congreso.backend.entities.SubmenuE;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubmenuRepo extends JpaRepository<SubmenuE,Long> {
+public interface SubmenuRepo extends JpaRepository<SubmenuE,Integer> {
 }
